Fixed bug in block disk cache.  It couldn't handle items with more than 127 blocks.  Now it can.  (Note:  for performance reasons, you should try to fit your items in as few blocks as possible.)
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@803209 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/jcs/auxiliary/disk/block/BlockDisk.java
@@ -177,7 +177,7 @@ public class BlockDisk
         int[] blocks = new int[numBlocksNeeded];
 
         // get them from the empty list or take the next one
-        for ( short i = 0; i < numBlocksNeeded; i++ )
+        for ( int i = 0; i < numBlocksNeeded; i++ )
         {
             Integer emptyBlock = (Integer) emptyBlocks.takeFirst();
             if ( emptyBlock != null )
@@ -194,7 +194,7 @@ public class BlockDisk
         byte[][] chunks = getBlockChunks( data, numBlocksNeeded );
 
         // write the blocks
-        for ( byte i = 0; i < numBlocksNeeded; i++ )
+        for ( int i = 0; i < numBlocksNeeded; i++ )
         {
             int position = calculateByteOffsetForBlock( blocks[i] );
             write( position, chunks[i] );
